Basename in install.vars
butler360 wrote:

I had to use that Mozilla line to get the interface to open, too, but
I was also seeing errors in the /opt/crashplan/bin/restart*.log for
missing Engine name. It also created so many restart logs I couldn't
delete them with rm, I had to use find. The file it referenced was
restartLinux.sh on line 21. That file led me to install.vars, where it
loads the variables. For some reason in the install.vars the
APP_BASENAME was commented out, which seems to have been causing the
restart error logs. I commented it back in and now it seems to be
working normally.